+++ hrmsAPI/methods/orders/serializers.py
@@ -1,9 +1,17 @@
 from rest_framework import serializers
-from ...models import Orders, OrderMenu
+from ...models import Orders, OrderMenu, Menu
 from ..tables.serializers import  TableSerializer
 from ..promocodes.serializers import  PromocodeSerializer
 from ..users.serializers import UserSerializer
-from ..orderMenu.serializers import OrderMenuSerializer
+from ..menu.serializers import MenuSerializer
+
+class OrderMenuSerializer(serializers.ModelSerializer):
+    
+    menu = MenuSerializer(read_only=True,many=True)
+
+    class Meta:
+        model=OrderMenu
+        fields=['id','order_id','menu','quantity']
 
 class OrderSerializer(serializers.ModelSerializer):
 
@@ -13,14 +21,16 @@ class OrderSerializer(serializers.ModelSerializer):
 
     promocode = PromocodeSerializer(read_only=True)
 
+    items = serializers.SerializerMethodField()
+
     class Meta:
         model = Orders
-        fields = ['id','user','table','promocode','created_at','complete_at']
+        fields = ['id','user','table','promocode','created_at','complete_at','items']
 
-    def to_representation(self, instance):
-        data = super().to_representation(instance)
-        data['items'] = OrderMenuSerializer(read_only=True,many=True).data
-        return data
+    def get_items(self, obj):
+        menus_ids = OrderMenu.objects.filter(order=obj).values_list('menu',flat=True)
+        menu = Menu.objects.filter(id__in=menus_ids)
+        return MenuSerializer(menu,many=True).data
 
     def create(self, validated_data):
         order_menu_data = validated_data.pop('order_menu')
